requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
idna==2.9
Pillow==6.2.1
python-dateutil==2.8.1
requests==2.23.0
six==1.15.0
urllib3==1.25.9
wluma-als-emulator==1.2.0
